app/requirements.txt
Flask==3.1.0
PyYAML==6.0.2
Requests==2.32.3
selenium==4.32.0
termcolor==3.1.0
Werkzeug==3.1.3